Split out FormulaCell
