+++ extension/source/cells/ExtensionLocalStorage.ts
@@ -2,11 +2,9 @@ import { EventEmitter } from 'events';
 
 import * as io from 'io-ts';
 import Browser from 'webextension-polyfill';
-import TypedEmitter from 'typed-emitter';
 
 import ExplicitAny from '../types/ExplicitAny';
-import AsyncReturnType from '../types/AsyncReturnType';
-import ICell, { IReadableCell, ReadableCellEmitter } from './ICell';
+import ICell, { ReadableCellEmitter } from './ICell';
 import CellIterator from './CellIterator';
 import jsonHasChanged from './jsonHasChanged';
 import assert from '../helpers/assert';
@@ -162,159 +160,4 @@ export class ExtensionLocalCell<T> implements ICell<T> {
       );
     }
   }
-}
-
-type InputValues<InputCells extends Record<string, IReadableCell<unknown>>> = {
-  [K in keyof InputCells]: AsyncReturnType<InputCells[K]['read']>;
-};
-
-export class FormulaCell<
-  InputCells extends Record<string, IReadableCell<unknown>>,
-  T,
-> implements IReadableCell<T>
-{
-  events = new EventEmitter() as ReadableCellEmitter<T>;
-
-  valuePromise: Promise<T>;
-  lastProvidedValue?: T;
-  ended = false;
-
-  constructor(
-    public inputCells: InputCells,
-    public formula: (inputValues: InputValues<InputCells>) => T,
-    public hasChanged = jsonHasChanged,
-  ) {
-    this.valuePromise = new Promise((resolve) => {
-      this.events.once('change', ({ latest }) => resolve(latest));
-    });
-
-    (async () => {
-      for await (const inputValues of toIterableOfRecords(inputCells)) {
-        if (this.ended) {
-          break;
-        }
-
-        const latest = formula(inputValues as InputValues<InputCells>);
-        this.valuePromise = Promise.resolve(latest);
-
-        if (hasChanged(this.lastProvidedValue, latest)) {
-          this.events.emit('change', {
-            previous: this.lastProvidedValue,
-            latest,
-          });
-        }
-      }
-
-      this.end();
-    })();
-  }
-
-  end() {
-    this.events.emit('end');
-    this.ended = true;
-  }
-
-  async read(): Promise<T> {
-    return await this.valuePromise;
-  }
-
-  [Symbol.asyncIterator](): AsyncIterator<T> {
-    return new CellIterator<T>(this);
-  }
-}
-
-/**
- * Wrapper around Object.keys which uses keyof to get the accurate key type.
- * (The builtin typing for Object.keys unnecessarily widens the type to
- * string[].)
- */
-function recordKeys<R extends Record<string, unknown>>(record: R): (keyof R)[] {
-  return Object.keys(record) as (keyof R)[];
-}
-
-type AsyncIteratee<I extends AsyncIterable<unknown>> = I extends AsyncIterable<
-  infer T
->
-  ? T
-  : never;
-
-function toIterableOfRecords<R extends Record<string, AsyncIterable<unknown>>>(
-  recordOfIterables: R,
-): AsyncIterable<{ [K in keyof R]: AsyncIteratee<R[K]> }> {
-  return {
-    [Symbol.asyncIterator]() {
-      const latest = {} as { [K in keyof R]: AsyncIteratee<R[K]> };
-      let providedFirstValue = false;
-      let keysFilled = 0;
-      const keysNeeded = recordKeys(recordOfIterables).length;
-      let ended = false;
-      let cleanup = () => {};
-
-      const events = new EventEmitter() as TypedEmitter<{
-        updated(): void;
-        end(): void;
-      }>;
-
-      function end() {
-        events.emit('end');
-        cleanup();
-        ended = true;
-      }
-
-      for (const key of recordKeys(recordOfIterables)) {
-        // eslint-disable-next-line no-loop-func
-        (async () => {
-          for await (const value of recordOfIterables[key]) {
-            if (ended) {
-              break;
-            }
-
-            if (!(key in latest)) {
-              keysFilled += 1;
-            }
-
-            latest[key] = value as ExplicitAny;
-
-            if (keysFilled === keysNeeded) {
-              events.emit('updated');
-            }
-          }
-        })();
-      }
-
-      return {
-        async next() {
-          if (!providedFirstValue && keysFilled === keysNeeded) {
-            providedFirstValue = true;
-            return { value: latest };
-          }
-
-          return new Promise<IteratorResult<typeof latest>>((resolve) => {
-            const updatedHandler = () => {
-              cleanup();
-              resolve({ value: latest });
-            };
-
-            events.on('updated', updatedHandler);
-
-            const endHandler = () => {
-              cleanup();
-              resolve({ value: undefined, done: true });
-            };
-
-            events.on('end', endHandler);
-
-            cleanup = () => {
-              events.off('updated', updatedHandler);
-              events.off('end', endHandler);
-            };
-          });
-        },
-        async return(): Promise<IteratorResult<typeof latest>> {
-          end();
-          return { value: undefined, done: true };
-        },
-      };
-    },
-  };
 }
